python/requirements.txt
cryptography
fastapi
python-jose
SQLAlchemy
uvicorn
psycopg[binary,pool] 
python-jose[cryptography] 
passlib[bcrypt] 
argon2-cffi 
python-multipart 
pydantic[email]
httpx